pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harmonium-lab"
version = "0.1.0"
description = "Music quality evaluation pipeline for harmonium procedural music engine"
requires-python = ">=3.11"
dependencies = [
    "music21>=9.0",
    "muspy>=0.5",
    "mido>=1.3",
    "pretty-midi>=0.2.10",
    "numpy>=1.24",
    "scipy>=1.10",
    "librosa>=0.10",
    "soundfile>=0.12",
]

[project.optional-dependencies]
audio = ["pyfluidsynth>=1.3"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
harmonium-lab = "harmonium_lab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/harmonium_lab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
